requirements.txt
PyYAML
requests
BeautifulSoup4
peewee
click
tk
logging
json
tldextract
datetime
collections